[2565] - REFACTOR: Move geolocation logic to model concern
